+++ v2/client/src/components/pages/Login/Login.style.js
@@ -40,33 +40,9 @@ export const LoginFail = styled.div`
   }
 `;
 
-export const Button = styled.div`
-  width: 100%;
-  line-height: 40px;
-  margin: 15px 0;
-  border-radius: 9px;
-  text-align: center;
-  font-weight: 400;
-  font-size: 1.4em;
-  border: 1px solid ${colors.primary};
-  background-color: ${colors.primary};
-  color: ${colors.white};
-  &:hover,
-  &:focus {
-    background-color: ${colors.white};
-    color: ${colors.primary};
-  }
-
-  @media (min-width: 768px) {
-    width: 100%;
-    margin: 0 auto;
-  }
-`;
-
 export const NoAccount = styled(InputDiv)`
   color: ${colors.black};
   margin-bottom: 0;
-
   &:not(:last-child) {
     margin-top: 6rem;
   }
@@ -77,8 +53,19 @@ export const NoAccount = styled(InputDiv)`
   }
 `;
 
-export const RegisterLink = styled(Link)`
+export const AnotherLink = styled(Link)`
   font-size: 1.1rem;
   font-weight: bold;
   color: ${colors.black};
+`;
+
+export const Content = styled.p`
+  text-align: center;
+  margin: 0 auto;
+
+  @media (min-width: 768px) {
+    width: 50%;
+    text-align: center;
+    margin: 2rem auto;
+  }
 `;
